LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/interfaces/WebSocket/events/020-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://example.invalid/' failed: Error resolving “example.invalid”: Name or service not known

PASS WebSockets: error events
